2007-03-01 14:10 UTC+0100 Przemyslaw Czerpak (druzus/at/priv.onet.pl)   * harbour/config/os2/global.cf   * harbour/contrib/hgf/os2pm/os2pm.c   * harbour/include/hbdefs.h   * harbour/include/hbgtcore.h   * harbour/source/rdd/dbcmd.c   * harbour/source/vm/mainpm.c     * cleaned OS2 build errors and warnings reported by David Macias.       Modifications not tested - OS2 users will have to make tests and       reports problems.

+++ harbour/contrib/hgf/os2pm/os2pm.c
@@ -291,8 +291,8 @@ HB_FUNC( WINSETOWNER )
 HB_FUNC( WINSENDMSG )
 {
    hb_retnl( ( LONG ) WinSendMsg( ( HWND ) hb_parnl( 1 ), hb_parnl( 2 ),
-             ( MPARAM ) IF( ISCHAR( 3 ), (ULONG) hb_parc( 3 ), hb_parnl( 3 ) ),
-             ( MPARAM ) IF( ISCHAR( 4 ), (ULONG) hb_parc( 4 ), hb_parnl( 4 ) ) ) );
+             ( MPARAM ) IF( ISCHAR( 3 ), (ULONG) hb_parc( 3 ), (ULONG) hb_parnl( 3 ) ),
+             ( MPARAM ) IF( ISCHAR( 4 ), (ULONG) hb_parc( 4 ), (ULONG) hb_parnl( 4 ) ) ) );
 }
